[#179] :sparkles: refactor: wonFormat 유틸성 함수 helper.ts 파일로 이동함에 따라 경로 수정
